CTCP-4276: refactoring app routes
