+++ app/controllers/representative/ActingAsRepresentativeController.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.representative
 
 import controllers.actions.Actions
 import forms.YesNoFormProvider
+import models.Mode
 import models.requests.MandatoryDataRequest
-import models.{LocalReferenceNumber, Mode}
 import navigation.Navigator
 import pages.ActingAsRepresentativePage
 import play.api.i18n.{I18nSupport, MessagesApi}
