+++ conf/app.routes
@@ -5,6 +5,7 @@
 ->         /                                                                  app.transitOperation.Routes
 ->         /                                                                  app.loading.Routes
 ->         /                                                                  app.transport.Routes
+->         /                                                                  app.representative.Routes
 
 ->         /hmrc-frontend                                                     hmrcfrontend.Routes
 GET        /assets/*file                                                      controllers.Assets.versioned(path = "/public", file: Asset)
@@ -36,8 +37,3 @@ GET        /:departureId                                                      co
 
 GET        /check-answers/:departureId                                        controllers.CheckYourAnswersController.onPageLoad(departureId: String)
 POST       /check-answers/:departureId                                        controllers.CheckYourAnswersController.onSubmit(departureId: String)
-
-GET        /representative/acting/:departureId                       controllers.ActingAsRepresentativeController.onPageLoad(departureId: String, mode: Mode = NormalMode)
-POST       /representative/acting/:departureId                       controllers.ActingAsRepresentativeController.onSubmit(departureId: String, mode: Mode = NormalMode)
-GET        /representative/change-acting/:departureId                controllers.ActingAsRepresentativeController.onPageLoad(departureId: String, mode: Mode = CheckMode)
-POST       /representative/change-acting/:departureId                controllers.ActingAsRepresentativeController.onSubmit(departureId: String, mode: Mode = CheckMode)
