+++ test/views/ActingAsRepresentativeViewSpec.scala
@@ -33,7 +33,7 @@ class ActingAsRepresentativeViewSpec extends YesNoViewBehaviours {
 
   behave like pageWithBackLink()
 
-  behave like pageWithSectionCaption("Trader details - Representative")
+  behave like pageWithSectionCaption("Representative")
 
   behave like pageWithHeading()
